backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-functions>=1.23.0",
    "azure-identity>=1.24.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "openai>=1.107.2",
    "pydantic>=2.11.7",
    "pymongo>=4.14.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "ruff>=0.12.11",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
